requirements.txt
fastapi
uvicorn
pydantic
redis
requests
google-generativeai # New dependency for Gemini
python-dotenv